dũng

+++ ss7-ss17_spring_boot/exercire/blog/src/main/java/com/codegym/blog/service/IGeneralService.java
@@ -5,9 +5,9 @@ import java.util.Optional;
 public interface IGeneralService<T> {
     Iterable<T> findAll();
 
-    Optional<T> findById(Long id);
+    Optional<T> findById(Integer id);
 
     void save(T t);
 
-    void remove(Long id);
+    void remove(Integer id);
 }
